Implemented town loading example, upgraded C# version
IMPORTANT: The C# language version is upgraded to 6 in this commit.
This allows the usage of “async” and “await” functionality, which is
extremely useful in asynchronous scene loading and transitions.

Other changes:
- Set up the GameManager system to work well with different game
“states.” Game states in this context include “showing book” and
“showing town.” There may be more later. This is implemented in the
GameManagerStates file using a state machine pattern.
- GameManagerHelpers contains helper functions for the GameManager
class.
- There are new entires in the GameManagerRegistry file. Three new
registering scripts are added: BookCameraScript, TownCamera and
TownEventSystem.
- The CameraFadeScript offers a comprehensive interface for putting a
fade-to-color effect on any camera.
- A button on the first page of the book causes a transition to an
example town scene.