Interaction system foundation
The InteractionSystem class keeps track of all interactive objects and
sends out events when they are added or removed. These events are currently
not used.

The IInteractiveObject interface defines the basic methods for an interactive
object:
- Can this object interact with a given interactor? Some objects may have distance constraints, for example.
- Try to interact with the interactor. This can fail sometimes.
- How far is this object from the interactor? This is used to figure out which object is closest for the purpose of selecting one.
- Show a visual to indicate that this object is selected.
- Hide the visual to indicate that this object is no longer selected.

The IInteractor class currently only provides a single GetPosition() method that can be used by interactive objects to compute their distance.

The InteractiveObjectTracker class helps efficiently find the closest interactive object to a particular interactor. This is a helper class that should be used inside IInteractor implementations.

TestInteractiveObject and TestInteractor are example scripts that implement basic functionality of the above interfaces. These can be added to objects inside a town scene without any extra work.