Added art, modified conditions system, created triggers
Trigger is an abstract class with an OnTrigger event that other code
can listen to. Trigger implementations use InvokeTrigger() to invoke
this method.

I removed trigger-like functionality from the ConditionManager. Also,
I moved the responsibility of activating reactions to the ReactionManager.

The hungry person in the Town1 scene now shows up properly.